requirements.txt
Flask==3.0.1
gunicorn==21.2.0
Jinja2==3.1.3
loguru==0.7.2
MarkupSafe==2.1.4
psutil==5.9.8
Werkzeug==3.0.1
python-dotenv==1.0.1